+++ src/main/java/org/generationcp/middleware/operation/builder/StandardVariableBuilder.java
@@ -176,18 +176,18 @@ public class StandardVariableBuilder extends Builder {
 	private void addConstraints(StandardVariable standardVariable, CVTerm cvTerm) throws MiddlewareQueryException {
 	    List<CVTermProperty> properties = getTermPropertyBuilder().findProperties(cvTerm.getCvTermId());
 		if (properties != null && !properties.isEmpty()) {
-            Integer minValue = null;
-            Integer maxValue = null;
+            Double minValue = null;
+            Double maxValue = null;
             Integer minValueId = null;
             Integer maxValueId = null;
 			
 		     for (CVTermProperty property : properties) {
 		         if (property.getTypeId().equals(TermId.MIN_VALUE.getId())){  
-                         minValue = Integer.parseInt(property.getValue());
+                         minValue = Double.parseDouble(property.getValue());
                          minValueId = property.getCvTermPropertyId();
 		         }
                  if (property.getTypeId().equals(TermId.MAX_VALUE.getId())){  
-                         maxValue = Integer.parseInt(property.getValue());
+                         maxValue = Double.parseDouble(property.getValue());
                          maxValueId = property.getCvTermPropertyId();
                  }
 		     }
